Migrate NavBar to composable and test
